Smart open to transfer datasets for universal transfer operator  (#1621)
# Description
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
Currently, all the files are being downloaded to temporary reference in
the worker and uploaded back to the destination dataset. This is
time-consuming and doesn't work well with database transfers. Hence,
implemented the transfer for all file types using smart open.

<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->
closes: #1619 


## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Add a feature to read the file using smart open and convert it into a
buffer that can be streamed into other data structures.
- Run the example DAG for S3 and GCS transfers.
- Add class for `Filetype` and `PandasDataframe`.
- For filetypes CSV, json and ndjson, read as `io.StringIO()`. For other
filetypes read as `io.BytesIO()`

## Does this introduce a breaking change?
No

### Checklist
- [x] Ran example DAG successfully.
- [ ] Extended the README / documentation, if necessary

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Phani Kumar <94376113+phanikumv@users.noreply.github.com>