Fix database to file issue (#1859)
# Description
## What is the current behavior?
Currently, when loading data from a database we are saving data in a
parquet file. Due to which the data when transferring data from the
database to the filesystem, we were limited to the parquet file type.

<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->


## What is the new behavior?
Now, we have to change the interface of read() also to give an output of
data frame and write() to consume data frame for csv, JSON, ndjson, and
parquet file type. This gives us the freedom to convert the data in any
format we prefer csv, Json, Ndjson or Parquet.

## Does this introduce a breaking change?
Nope

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

---------

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>