Add logic to convert dict to Options class (#55)
# Description
## What is the current behavior?
Currently, for transfer_params we have a logic to convert dict to
TransferIntegrationOptions but that doesn't seem to be executed. Due to
this, we are essentially using dict type in our codebase assuming it's
of type TransferIntegrationOptions. Which seems to mess with static type
check.


closes:
https://github.com/astronomer/apache-airflow-provider-transfers/issues/51

## What is the new behavior?
We need to use the TransferIntegrationOptions class or its subclass
internally, so any dict that is passed to transfer_params should be
converted to TransferIntegrationOptions or a subclass.

Algorithm:

1. If users have passed a `TransferIntegrationOptions` or subclass of
it. We let it pass, assuming the user knows what he is doing.
2. If the user have passed a `dict`, convert it to an Option class, We
check if the transfer mode is third-party,
1. Third-party - Then the logic for getting the correct `Ingestor` class
comes from `conn_id` passed in `transfer_param`. Once we get the
`Ingestor` class we look for the `OPTIONS_CLASS` attribute, this
attribute is supposed to hold a reference to correct `Options Class`
         1. Attribute found - We dict to instantiate this class.
2. Attribute not found - We use the class `TransferIntegrationOptions`
and instantiate it with dict.
2. Native - Then the logic for getting the correct options class goes to
`provider` of `destination_dataset`. we determine`provider` class by
`conn_id` of `destination_dataset`. Once we have the `provider` class,
we look for `OPTIONS_CLASS` attribute, this attribute is supposed to
hold a reference to correct `OPTIONS_CLASS`
1. `OPTIONS_CLASS` Attribute found - We use dict to instantiate this
class.
2. `OPTIONS_CLASS` Attribute not found - We use the class
`TransferIntegrationOptions` and instantiate it with dict.
       
## Does this introduce a breaking change?
Nope


### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary